delete IT tests and added more unit tests
