+++ .pdd
@@ -2,6 +2,7 @@
 --verbose
 --exclude target/**/*
 --exclude src/site/resources/**/*
+--exclude src/test/resources/org/cactoos/http/keystore
 --rule min-words:20
 --rule min-estimate:15
 --rule max-estimate:90
